Completed labelled array datatype and refractored datatypes
